Fix rendering logic
I was trying to be too clever by half.  WikiMedia loads the js after the
DOM is loaded so it will be safe to just run the map logic directly
instead of trying to link it to a window load event.

The need for this change is that the previous logic had a bug where the
map rendering wouldn't load in certain cases (e.g. if the page content
was rendered after saving the source).

Issue #7